requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.1.0
constantly==23.10.4
coverage==7.6.10
cryptography==42.0.8
daphne==4.1.2
Django==4.2.3
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dotenv==1.1.1
redis==5.2.1
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
zope.interface==7.2
